Programming-WebApp/todolist-project/requirements.txt
asgiref==3.8.1
Django==4.2.21
djangorestframework==3.16.1
sqlparse==0.5.3
typing_extensions==4.13.2
